Only allow integer number font sizes for simplicity

+++ include/font.h
@@ -5,11 +5,12 @@
 #include <SDL3/SDL_gpu.h>
 #include <SDL3/SDL_iostream.h>
 #include <SDL3/SDL_pixels.h>
+#include <SDL3/SDL_stdinc.h>
 
 typedef struct font_t font_t;
 
 [[nodiscard]]
-font_t *font_create(SDL_GPUDevice *device, SDL_IOStream *source, float font_size, SDL_Color color);
+font_t *font_create(SDL_GPUDevice *device, SDL_IOStream *source, Uint16 font_size, SDL_Color color);
 
 void font_destroy(font_t *font);
